Hooked up Auth store to login page
The authentication store is now hooked up to the login page. When the application is ran, it will start on the login page. When you press login, it will bring you to the nav bar view, regardless of user input. User and admin logins are NOT being validated with the database. This commit is to have the store in a working state and help with the flow of the application.