Redirect user to login
User will be redirected to the login page if they try and go to any other page. The 'logout' button in the navbar will now log out the user. If a page is reloaded, the user will be logged out. I Need to add a state manager to keep the authorized state of the user on a page reload and reroute to '/' on successful login.